crates/models: reorganize existing model layout within `names` module
Issue #267
