+++ crates/models/src/names/schema_support.rs
@@ -1,7 +1,3 @@
-use schemars::schema;
-use serde_json::{from_value, json};
-use std::time::Duration;
-
 use crate::names::*;
 
 impl Lambda {
@@ -11,88 +7,4 @@ impl Lambda {
     pub fn example_remote() -> Self {
         Self::Remote("http://example/api".to_string())
     }
-}
-
-impl PartitionSelector {
-    pub fn example() -> Self {
-        from_value(json!({
-            "include": {
-                "a_partition": ["A", "B"],
-            },
-            "exclude": {
-                "other_partition": [32, 64],
-            }
-        }))
-        .unwrap()
-    }
-}
-
-impl CompressionCodec {
-    pub fn example() -> Self {
-        CompressionCodec::GzipOffloadDecompression
-    }
-}
-
-impl BucketType {
-    pub fn example() -> Self {
-        BucketType::S3
-    }
-}
-
-impl Store {
-    pub fn example() -> Self {
-        Self {
-            provider: BucketType::S3,
-            bucket: "my-bucket".to_string(),
-            prefix: None,
-        }
-    }
-}
-
-impl StorageMapping {
-    pub fn example() -> Self {
-        Self {
-            prefix: Prefix::new("acmeCo/widgets/"),
-            stores: vec![Store::example()],
-        }
-    }
-}
-
-impl FragmentTemplate {
-    pub fn example() -> Self {
-        Self {
-            compression_codec: Some(CompressionCodec::Zstandard),
-            flush_interval: Some(Duration::from_secs(3600)),
-            ..Default::default()
-        }
-    }
-    pub fn duration_schema(g: &mut schemars::gen::SchemaGenerator) -> schema::Schema {
-        duration_schema(g)
-    }
-}
-
-impl JournalTemplate {
-    pub fn example() -> Self {
-        Self {
-            fragments: FragmentTemplate::example(),
-        }
-    }
-}
-
-impl ShardTemplate {
-    pub fn example() -> Self {
-        Self {
-            max_txn_duration: Some(Duration::from_secs(30)),
-            hot_standbys: Some(1),
-            ..Default::default()
-        }
-    }
-}
-
-fn duration_schema(_: &mut schemars::gen::SchemaGenerator) -> schema::Schema {
-    from_value(json!({
-        "type": ["string", "null"],
-        "pattern": "^\\d+(s|m|h)$"
-    }))
-    .unwrap()
 }
